[WIP] started migrating containers using resfreshableFeed in their components to use reselect
