+++ package.json
@@ -69,6 +69,7 @@
     "redux": "=4.0.1",
     "redux-persist": "=5.10.0",
     "redux-thunk": "=2.3.0",
+    "reselect": "^4.0.0",
     "stream-browserify": "=2.0.1",
     "string_decoder": "=1.1.1",
     "ts-jest": "=23.10.4",
