+++ src/containers/FavoritesContainer.ts
@@ -2,21 +2,13 @@ import { connect } from 'react-redux';
 import { AppState } from '../reducers';
 import { StateProps, DispatchProps } from '../components/FavoritesFeedView';
 import { AsyncActions } from '../actions/Actions';
-import { Post } from '../models/Post';
 import { Feed } from '../models/Feed';
 import { FavoritesFeedView } from '../components/FavoritesFeedView';
-
-const isPostFromFavoriteFeed = (post: Post, favoriteFeeds: Feed[]): boolean => {
-    return favoriteFeeds.find(feed => {
-        return feed != null && post.author != null &&
-            feed.feedUrl === post.author.uri;
-    }) != null;
-};
+import { getFavoriteFeedsPosts, getFavoriteFeeds } from '../selectors/selectors';
 
 const mapStateToProps = (state: AppState, ownProps): StateProps => {
-    const favoriteFeeds = state.feeds.filter(feed => feed != null && feed.favorite === true);
-    const posts = state.rssPosts
-        .filter(post => post != null && isPostFromFavoriteFeed(post, favoriteFeeds));
+    const posts = getFavoriteFeedsPosts(state);
+    const favoriteFeeds = getFavoriteFeeds(state);
 
     return {
         navigation: ownProps.navigation,
